feat: add token usage display in RuntimePanel and session deletion functionality
- Enhanced RuntimePanel to display token usage details including calls, input tokens, output tokens, total tokens, and estimated calls.
- Updated SessionSidebar to include a delete button for sessions, with a confirmation prompt before deletion.
- Implemented deleteSession API function to handle session deletion on the backend.
- Modified WorkspacePage to manage session deletion and update session state accordingly.
- Improved CSS styles for session items and delete button for better UI/UX.
- Updated TypeScript types to include optional token usage details in RuntimeInfo.

+++ backend/src/interfaces/api/rate_limit.py
@@ -88,5 +88,5 @@ async def enforce_chat_rate_limit(request: Request) -> None:
     raise HTTPException(
         status_code=429,
         detail="Rate limit exceeded for /chat",
-        headers={"Retry-After": str(int(math.ceil(retry_after)))},
+        headers={"Retry-After": str(math.ceil(retry_after))},
     )
